requirements.txt
streamlit>=1.30.0
google-generativeai>=0.4.0
Pillow>=10.0.0
scikit-learn>=1.2.2
numpy>=1.24.0
pandas>=2.0.0
shap>=0.44.0
matplotlib>=3.7.0